requirements.txt
pandas>=1.5.0
sqlalchemy>=1.4.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
streamlit>=1.28.0
plotly>=5.15.0